Bug fixes: Date sync between lucene index and db (hacked). Make study public when resubmitting Do not increase the accession counter when resubmitting.
Queue system:
Stability changes (tested intensively with an empty db and submitting, resubmitting, changing public release date and checking locations of files and folder during the whole process)

Config page: new tab for checking study health (if it is placed in the appropriate folder and moved when made public).

git-svn-id: https://svn.code.sf.net/p/metabolomes/code/trunk@1070 71bfadd5-8d58-49b2-b2b0-ff0b8e09f556

+++ metabolights-webapps/src/main/sql/EmptyDB.oracle.sql
@@ -1,4 +1,4 @@
---WARNING: This will remove the users and the stable id
+--WARNING: This will NOT remove the users and the stable id (Comment the IF if you want to do so).
 
 declare
 
@@ -16,9 +16,11 @@ begin
 
  for cur2 in c2 loop
    begin
-     execute immediate 'ALTER TABLE ISATAB.'||cur2.table_name||' disable all triggers';
-     execute immediate 'TRUNCATE TABLE ISATAB.'||cur2.table_name;
-     execute immediate 'ALTER TABLE ISATAB.'||cur2.table_name||' enable all triggers';
+	 IF (cur2.table_name <> 'USER_DETAIL' AND cur2.table_name <> 'STABLE_ID') THEN  
+     	execute immediate 'ALTER TABLE ISATAB.'||cur2.table_name||' disable all triggers';
+     	execute immediate 'TRUNCATE TABLE ISATAB.'||cur2.table_name;
+     	execute immediate 'ALTER TABLE ISATAB.'||cur2.table_name||' enable all triggers';
+     END IF;
    exception when others then null;
   end;
  end loop;
